requirements.txt
fastapi
uvicorn
pytesseract
pdfplumber
Pillow
python-docx
requests
sentence-transformers
faiss-cpu
